Refactored the dev control inputs It now lives in debug_tools
