+++ emergence_lib/src/hive_mind.rs
@@ -5,7 +5,6 @@ use crate::signals::emitters::Emitter;
 use crate::signals::emitters::StockEmitter::{PheromoneAttract, PheromoneRepulse};
 use crate::signals::SignalModificationEvent;
 use bevy::prelude::*;
-use debug_tools::{DebugInfo, DebugToolsPlugin};
 use leafwing_input_manager::prelude::*;
 
 /// Provides the interface between the player and the hive.
@@ -14,12 +13,8 @@ pub struct HiveMindPlugin;
 impl Plugin for HiveMindPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugin(InputManagerPlugin::<HiveMindAction>::default())
-            .add_plugin(InputManagerPlugin::<DevAction>::default())
-            .init_resource::<DebugInfo>()
-            .add_plugin(DebugToolsPlugin)
             .add_startup_system(initialize_hive_mind)
-            .add_system(place_pheromone)
-            .add_system(show_debug_info);
+            .add_system(place_pheromone);
     }
 }
 
@@ -34,20 +29,6 @@ pub enum HiveMindAction {
     PlaceAttractivePheromone,
     /// Place a repulsive pheromone.
     PlaceRepulsivePheromone,
-}
-
-/// Enumerates the actions a developer can take.
-#[derive(Actionlike, Clone, Copy, PartialEq, Eq, Debug)]
-pub enum DevAction {
-    /// Toggle the overall developer mode setting
-    ToggleDevMode,
-    // TODO: make debug labels
-    /// Toggle tilemap labels tools
-    ToggleTileLabels,
-    /// Toggle rendering info
-    ToggleInfoText,
-    /// Toggle the inspector
-    ToggleInspector,
 }
 
 // TODO: rework this to use LWIM conventions for mapping controls
@@ -69,34 +50,9 @@ impl Default for HiveMindControls {
     }
 }
 
-/// Interface for developer controls
-pub struct DevControls {
-    /// Toggle the dev mode
-    pub toggle_dev_mode: UserInput,
-    /// Toggle the tile label
-    pub toggle_tile_labels: UserInput,
-    /// Toggle the fps ui
-    pub toggle_fps: UserInput,
-    /// Toggle the inspector
-    pub toggle_inspector: UserInput,
-}
-
-/// Add default developer controls
-impl Default for DevControls {
-    fn default() -> Self {
-        Self {
-            toggle_dev_mode: UserInput::chord([KeyCode::LControl, KeyCode::LShift, KeyCode::D]),
-            toggle_tile_labels: UserInput::chord([KeyCode::LControl, KeyCode::LShift, KeyCode::T]),
-            toggle_fps: UserInput::chord([KeyCode::LControl, KeyCode::LShift, KeyCode::V]),
-            toggle_inspector: UserInput::chord([KeyCode::LControl, KeyCode::LShift, KeyCode::I]),
-        }
-    }
-}
-
 /// Startup system initializing the [`HiveMind`].
 fn initialize_hive_mind(mut commands: Commands) {
     let controls = HiveMindControls::default();
-    let dev_controls = DevControls::default();
     commands
         .spawn_empty()
         .insert(HiveMind)
@@ -111,15 +67,6 @@ fn initialize_hive_mind(mut commands: Commands) {
                     controls.repulsive_pheromone,
                     HiveMindAction::PlaceRepulsivePheromone,
                 ),
-            ]),
-        })
-        .insert(InputManagerBundle::<DevAction> {
-            action_state: ActionState::default(),
-            input_map: InputMap::new([
-                (dev_controls.toggle_dev_mode, DevAction::ToggleDevMode),
-                (dev_controls.toggle_tile_labels, DevAction::ToggleTileLabels),
-                (dev_controls.toggle_fps, DevAction::ToggleInfoText),
-                (dev_controls.toggle_inspector, DevAction::ToggleInspector),
             ]),
         });
 }
@@ -151,53 +98,5 @@ fn place_pheromone(
             pos: (*cursor_tile_pos).unwrap(),
             increment: 0.01,
         });
-    }
-}
-
-/// Toggle showing debug info
-fn show_debug_info(
-    dev: Query<&ActionState<DevAction>, With<HiveMind>>,
-    mut debug_info: ResMut<DebugInfo>,
-) {
-    let dev = dev.single();
-    let dev_mode = dev.just_pressed(DevAction::ToggleDevMode);
-
-    // Toggle the dev mode so that what happens is intuitive to the user
-    if dev_mode {
-        if debug_info.dev_mode {
-            debug_info.disable();
-            info!("Debug Info disabled");
-        } else {
-            debug_info.enable();
-            info!("Debug Info enabled");
-        }
-    }
-
-    if !debug_info.dev_mode {
-        // There is no other work to be done on this callback
-        // The DebugInfo shouldn't be changed after it gets disabled, or is already disabled
-        return;
-    }
-
-    let tile_labels = dev.just_pressed(DevAction::ToggleTileLabels);
-    let fps_info = dev.just_pressed(DevAction::ToggleInfoText);
-    let inspector = dev.just_pressed(DevAction::ToggleInspector);
-
-    toggle_debug_var(tile_labels, &mut debug_info.show_tile_labels, "Tile labels");
-    toggle_debug_var(fps_info, &mut debug_info.show_fps_info, "FPS info");
-    toggle_debug_var(inspector, &mut debug_info.show_inspector, "Egui inspector");
-}
-
-fn toggle_debug_var(is_action_active: bool, var: &mut bool, log_str: &'static str) {
-    if !is_action_active {
-        return;
-    }
-
-    if *var {
-        *var = false;
-        info!("{} off", log_str);
-    } else {
-        *var = true;
-        info!("{} on", log_str);
     }
 }
