+++ emergence_lib/src/lib.rs
@@ -25,7 +25,8 @@ impl Plugin for InteractionPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugin(camera::CameraPlugin)
             .add_plugin(cursor::CursorTilePosPlugin)
-            .add_plugin(hive_mind::HiveMindPlugin);
+            .add_plugin(hive_mind::HiveMindPlugin)
+            .add_plugin(debug_tools::DebugToolsPlugin);
     }
 }
